🔧 Fix: Change uint to uint256 for type consistency

+++ src/CustomErrors.sol
@@ -3,16 +3,16 @@ pragma solidity 0.8.28;
 
 contract CustomErrors {
     address public owner;
-    uint public number;
+    uint256 public number;
 
     error NotOwner(address caller);
-    error TooLow(uint sent, uint required);
+    error TooLow(uint256 sent, uint256 required);
 
     constructor() {
         owner = msg.sender;
     }
 
-    function setNumber(uint value) public {
+    function setNumber(uint256 value) public {
         if (msg.sender != owner) {
             revert NotOwner(msg.sender);
         } else if (value < 10) {
